Added: New method in Util class for checking the password strength ( Util.passwordStrength(String) )
The function does a simple regex find to locate any occurance of a letter of any language, number and any punctuation or whitespace.
The 3 groups are added up together and the result is the strength value of the password ranging between 1-3.
If the password consisting of any of these 3 groups is less than 8 characters in length, return value is 0.